fix: delete thead, just make getter instead
