+++ packages/components/src/components.ts
@@ -23,7 +23,6 @@ export { default as HdsAdvancedTableThButtonTooltip } from './components/hds/adv
 export { default as HdsAdvancedTableThSelectable } from './components/hds/advanced-table/th-selectable.ts';
 export { default as HdsAdvancedTableThSort } from './components/hds/advanced-table/th-sort.ts';
 export { default as HdsAdvancedTableTr } from './components/hds/advanced-table/tr.ts';
-export { default as HdsAdvancedTableThead } from './components/hds/advanced-table/thead.ts';
 export { default as HdsAdvancedTableThButtonExpand } from './components/hds/advanced-table/th-button-expand.ts';
 export { default as HdsAdvancedTableExpandableTrGroup } from './components/hds/advanced-table/expandable-tr-group.ts';
 export * from './components/hds/advanced-table/types.ts';
