--- packages/components/src/components/hds/advanced-table/thead.hbs
@@ -1,7 +1,0 @@
-{{!
-  Copyright (c) HashiCorp, Inc.
-  SPDX-License-Identifier: MPL-2.0
-}}
-<div class={{this.classNames}} role="rowgroup">
-  {{yield}}
-</div>
